Update sm2135.cpp
Fixes all issues I could find with this code. Changed the CW address to match example code, added specific currents to CW and RGB, changed the update_channel conditions and added a delay after the commands to give time for ACK. In my own testing everything now works as long as the color and cwww controls are separated.